feat : cart

+++ database/migrations/2024_02_12_070333_create_headers_detail_customers_carts_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
             $table->integer('quantity');
-            $table->integer('sub_total');
+            $table->float('sub_total');
             $table->timestamps();
         });
     }
